fwk: Fix mf propagation when mux demux is placed at boundary of a subgraph
When mux_demux is placed at boundary of subgraph and downstream subgraph
is closed and reopened, the media format was not getting propagated.
This fix ensures that the output media format sum is reset when the port
is closed, and the media format is raised again when the port is
reopened.

Signed-off-by: Aditya Rathi <quic_aditrath@quicinc.com>